Enable deep storage custom right click menu config
useRightClickLogic is now true!

This will allow you to right click directly on say a gun cabinent to
access "equip to side-arm" menu for a pawn. Does not cause desync w/
arbiter in local testing. Otherwise as the drop to floor eject button is
off, it is difficult to access a specific item in a deep storage stack.